[Project structure] KMM core module
* separated plugins from libs from versions
