+++ app/build.gradle.kts
@@ -1,7 +1,7 @@
-import Config.androidBuildTools
-import Config.androidCompileSdk
-import Config.androidMinSdk
-import Config.androidTargetSdk
+import ApplicationConfig.androidBuildTools
+import ApplicationConfig.androidCompileSdk
+import ApplicationConfig.androidMinSdk
+import ApplicationConfig.androidTargetSdk
 
 plugins {
     id(Plugins.appPlugin)
