+++ buildSrc/src/main/java/ApplicationConfig.kt
@@ -1,0 +1,10 @@
+object ApplicationConfig {
+    // Android config
+    const val androidBuildTools = "31.0.0"
+    const val androidMinSdk = 21
+    const val androidTargetSdk = 31
+    const val androidCompileSdk = 31
+    const val applicationId = "com.github.mik629.android.fundamentals"
+    const val versionCode = 1
+    const val versionName = "1.0"
+}
