+++ buildSrc/src/main/java/Plugins.kt
@@ -1,0 +1,10 @@
+object Plugins {
+    const val androidPlugin = "android"
+    const val appPlugin = "com.android.application"
+    const val kapt = "kapt"
+    const val multiplatform = "multiplatform"
+    const val hiltPlugin = "dagger.hilt.android.plugin"
+    const val serializationPlugin = "plugin.serialization"
+    const val versions = "com.github.ben-manes.versions"
+    const val library = "com.android.library"
+}
