+++ buildSrc/src/main/java/Versions.kt
@@ -1,0 +1,31 @@
+object Versions {
+    const val kotlin = "1.5.31"
+    const val activityCompose = "1.4.0"
+    const val compose = "1.0.5"
+    const val vmCompose = "2.4.0-rc01"
+    const val navigationCompose = "2.4.0-alpha10"
+    const val hiltNavigationCompose = "1.0.0-rc01"
+    const val lifecycle = "2.4.0"
+    const val workManager = "2.7.1"
+
+    // Plugins
+    const val androidToolsPlugin = "7.0.4"
+    const val versionsPlugin = "0.41.0"
+
+    // Android libraries
+    const val androidCore = "1.7.0"
+    const val coroutines = "1.6.0"
+    const val dagger = "2.40.5"
+    const val dataStore = "1.0.0"
+    const val playServicesAuth = "20.0.1"
+    const val serialization = "1.3.2"
+
+    // third party libs
+    const val glideCompose = "1.4.5"
+    const val timber = "5.0.1"
+
+    // testing
+    const val junit = "4.13.2"
+    const val junitExt = "1.1.3"
+    const val kaspresso = "1.4.0"
+}
